+++ test/runtests.jl
@@ -10,8 +10,10 @@ using SUNDMRG
     include("test_tables_small.jl")
     include("test_tables_ground_truth.jl")
     include("test_representation_theory_internal.jl")
+    include("test_representation_theory_definitions.jl")
     include("test_step_helpers.jl")
     include("test_lanczos_helpers.jl")
     include("test_tools_onthefly.jl")
+    include("test_wigner_racah_definitions.jl")
     include("test_run_dmrg.jl")
 end